+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Features
+
+- Profiles are available via the `--profile` command line argument. If omitted,
+  the default profile will be used, which is the same as using
+  `--profile default`.
+
 ## 2.1.0 - 2022-09-15
 
 ### Features